requirements.txt
asgiref==3.6.0
Django==4.1.4
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
gunicorn==20.1.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
pytz==2022.7
sqlparse==0.4.3
